experimentations/trustworthiness/fairness/holisticAI/requirements.txt
lightgbm
fairlearn
xlrd
seaborn
imbalanced-learn==0.13.0
sklearn-compat==0.1.3
scikit-learn==1.6.1
holisticai==1.0.14
gdown
jax
optax
flax
ipywidgets

pyyaml
interpret
aix360
lime